requirements.txt
# ============================================================
# 공통 라이브러리 (모든 에이전트에서 사용)
# ============================================================
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
setuptools==69.5.1      # pykrx가 의존하는 pkg_resources 제공 (Python 3.13 호환)


# ============================================================
# 섹터/종목 에이전트 (sector/)
# ============================================================
pykrx>=1.0.45           # 한국 주식 OHLCV, 수급, 밸류에이션 (수급·상대강도·밸류에이션 모듈)
beautifulsoup4>=4.12.0  # 네이버 증권 크롤링 (naver_finance 모듈)
lxml>=4.9.0             # BeautifulSoup HTML 파서 백엔드


# ============================================================
# 불/베어 에이전트 (bull-bear/)
# ============================================================
# pykrx — 위 섹터 항목과 공유
pandas_ta>=0.3.14b      # 기술적 지표 계산 (RSI, MACD, 볼린저 밴드)

# ============================================================
# 매크로 에이전트 (macro/)
# ============================================================
scikit-learn>=1.3.0
statsmodels>=0.14.0
